+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Yii Message Translator Change Log
 
-## 2.2.2 under development
+## 3.0.0 under development
 
 - Chg #92: Implement fluent interface into `TranslatorInterface`: methods `addCategorySources()`, `setLocale()`,
   `withDefaultCategory()` and `withLocale()` returns `static` (@vjik)
+- Chg #111: Adapt configuration group names to Yii conventions (@vjik)
 
 ## 2.2.1 January 07, 2023